# openapi.json
{
  "components": {
    "schemas": {
...
      "DownloadListResponse": {
        "properties": {
          "downloads": {
            "items": { "$ref": "#/components/schemas/DownloadResponse" },
            "title": "Downloads",
            "type": "array"
          },
          "pagination": { "$ref": "#/components/schemas/PaginationInfo" }
        },
        "required": ["downloads", "pagination"],
        "title": "DownloadListResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Standard error response schema.",
        "example": {
          "details": { "job_id": "123" },
          "error": { "code": "NOT_FOUND", "message": "Resource not found" }
        },
        "properties": {
          "details": {
            "anyOf": [{ "additionalProperties": true, "type": "object" }, { "type": "null" }],
            "title": "Details"
          },
          "error": {
            "additionalProperties": { "type": "string" },
            "title": "Error",
            "type": "object"
          }
        },
        "required": ["error"],
        "title": "ErrorResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": { "HTTPBearer": { "scheme": "bearer", "type": "http" } }
  },
  "info": {
    "contact": { "name": "Team 21", "url": "https://github.com/tomkabel/team21-vooglaadija" },
    "description": "REST API for user authentication, creating download jobs, tracking job status, and retrieving processed files. Authentication uses bearer JWT access tokens.",
    "license": { "name": "GPLv3", "url": "https://www.gnu.org/licenses/gpl-3.0.html" },
    "summary": "Asynchronous API for authenticated video download jobs.",
    "title": "Vooglaadija API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/downloads": {
      "get": {
        "description": "Return paginated download jobs belonging to the authenticated user.",
        "operationId": "list_downloads_api_v1_downloads_get",
        "parameters": [
          {
            "description": "Page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page",
              "maximum": 100,
              "minimum": 1,
              "title": "Per Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "downloads": [
                    {
                      "completed_at": "2026-04-07T12:01:00Z",
                      "created_at": "2026-04-07T12:00:00Z",
                      "expires_at": "2026-04-08T12:01:00Z",
                      "file_name": "video.mp4",
                      "id": "550e8400-e29b-41d4-a716-446655440000",
                      "status": "completed",
                      "url": "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
                    }
                  ],
                  "pagination": { "page": 1, "per_page": 20, "total": 1 }
                },
                "schema": { "$ref": "#/components/schemas/DownloadListResponse" }
              }
            },
            "description": "List of download jobs"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": { "code": "UNAUTHORIZED", "message": "Could not validate credentials" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "details": {
                    "validation_errors": [
                      {
                        "field": "query.page",
                        "message": "Input should be greater than or equal to 1",
                        "type": "greater_than_equal"
                      }
                    ]
                  },
                  "error": { "code": "VALIDATION_ERROR", "message": "Request validation failed" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Validation error"
          }
        },
        "security": [{ "HTTPBearer": [] }],
        "summary": "List download jobs",
        "tags": ["downloads"]
      },
      "post": {
        "description": "Queue a new YouTube download job for the authenticated user.",
        "operationId": "create_download_api_v1_downloads_post",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/DownloadCreate" } }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "created_at": "2026-04-07T12:00:00Z",
                  "id": "550e8400-e29b-41d4-a716-446655440000",
                  "status": "pending",
                  "url": "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
                },
                "schema": { "$ref": "#/components/schemas/DownloadResponse" }
              }
            },
            "description": "Download job created"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": { "code": "UNAUTHORIZED", "message": "Could not validate credentials" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "details": {
                    "validation_errors": [
                      {
                        "field": "url",
                        "message": "Value error, Must be a valid YouTube URL",
                        "type": "value_error"
                      }
                    ]
                  },
                  "error": { "code": "VALIDATION_ERROR", "message": "Request validation failed" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Validation error"
          }
        },
        "security": [{ "HTTPBearer": [] }],
        "summary": "Create download job",
        "tags": ["downloads"]
      }
    },
    "/api/v1/downloads/{job_id}": {
      "delete": {
        "description": "Delete a download job and remove its file from storage when present.",
        "operationId": "delete_download_api_v1_downloads__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": { "title": "Job Id", "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Download job deleted" },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": { "code": "UNAUTHORIZED", "message": "Could not validate credentials" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "details": { "job_id": "unknown-id" },
                  "error": { "code": "NOT_FOUND", "message": "Download job not found" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Download job not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [{ "HTTPBearer": [] }],
        "summary": "Delete download job",
        "tags": ["downloads"]
      },
      "get": {
        "description": "Return a specific download job by id if it belongs to the current user.",
        "operationId": "get_download_api_v1_downloads__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": { "title": "Job Id", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "created_at": "2026-04-07T12:00:00Z",
                  "id": "550e8400-e29b-41d4-a716-446655440000",
                  "status": "processing",
                  "url": "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
                },
                "schema": { "$ref": "#/components/schemas/DownloadResponse" }
              }
            },
            "description": "Download job details"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": { "code": "UNAUTHORIZED", "message": "Could not validate credentials" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "details": { "job_id": "unknown-id" },
                  "error": { "code": "NOT_FOUND", "message": "Download job not found" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Download job not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [{ "HTTPBearer": [] }],
        "summary": "Get download job",
        "tags": ["downloads"]
      }
    },
    "/api/v1/downloads/{job_id}/file": {
      "get": {
        "description": "Download the processed file for a completed, non-expired download job.",
        "operationId": "get_download_file_api_v1_downloads__job_id__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": { "title": "Job Id", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": { "application/json": { "schema": {} } },
            "description": "Binary file stream"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Job is not completed. Current status: processing"
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Job not completed"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": { "code": "UNAUTHORIZED", "message": "Could not validate credentials" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "error": { "code": "FORBIDDEN", "message": "Access denied: invalid file path" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Invalid file path"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "details": { "job_id": "550e8400-e29b-41d4-a716-446655440000" },
                  "error": { "code": "NOT_FOUND", "message": "File not found" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Job or file not found"
          },
          "410": {
            "content": {
              "application/json": {
                "example": {
                  "error": { "code": "VALIDATION_ERROR", "message": "Download link has expired" }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Download link expired"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [{ "HTTPBearer": [] }],
        "summary": "Download output file",
        "tags": ["downloads"]
      }
    },
    "/api/v1/downloads/{job_id}/retry": {
      "post": {
        "description": "Retry a failed download job.",
        "operationId": "retry_download_api_v1_downloads__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": { "title": "Job Id", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/DownloadResponse" } }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [{ "HTTPBearer": [] }],
        "summary": "Retry Download",
        "tags": ["downloads"]
      }
    },
...
  },
...
}